refactor: split integration tests into per-topic files
Replace the single tests/integration.rs (with section separators) with
one file per topic: send_traces, send_stats, send_pipeline_stats,
send_telemetry, send_evp_event, agent_info, static_headers. Shared
helpers (ensure_crypto_provider, client_for) live in tests/common/mod.rs.
